Adding in official sonnet name
Reason:
The response streaming chunks only return with the name of this exact model name, with the date, even though we called it with "claude-3-7-sonnet-latest". 

But because of this PromptingTools.get(PromptingTools.MODEL_REGISTRY, model, (; cost_of_token_prompt=0.0, cost_of_token_generation=0.0))
codes are like this are silently failing, with returning zeros for prices.